+++ simpleble/include/simpleble_c/types.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <stddef.h>
+#include <stdbool.h>
 
 #define SIMPLEBLE_UUID_STR_LEN 37  // 36 characters + null terminator
 #define SIMPLEBLE_CHARACTERISTIC_MAX_COUNT 16
@@ -22,6 +23,11 @@ typedef struct {
 
 typedef struct {
     simpleble_uuid_t uuid;
+    bool can_read;
+    bool can_write_request;
+    bool can_write_command;
+    bool can_notify;
+    bool can_indicate;
     size_t descriptor_count;
     simpleble_descriptor_t descriptors[SIMPLEBLE_DESCRIPTOR_MAX_COUNT];
 } simpleble_characteristic_t;